fix(compiler): Removes several warnings related to sign
Fix -Wsign-compare warnings from int/size_t mismatches in loops from Windows compiler messages

- Change loop variables from int to size_t in fims_math.hpp (2 locations)
- Change loop variables from int to size_t in rcpp_models.hpp (3 locations)
- Change loop variables from int to size_t in rcpp_population.hpp (1 location)
- Change loop variables from int to size_t in rcpp_distribution.hpp (3 locations)
- Add static_cast<size_t> for n_years.get() comparison in rcpp_recruitment.hpp
- Fix uint32_t signedness warning in rcpp_growth.hpp
- Change loop variable from uint32_t to size_t for consistency with .size() return type
- Fix R_xlen_t vs size_t signedness warnings
- Change loop variables to match container type
- Use R_xlen_t for Rcpp vectors (Rcpp::NumericVector, Rcpp::IntegerVector, Rcpp::List)
- Use size_t for C++ containers (std::vector, RealVector, ParameterVector)
- Use .assign() and static_cast to fix signedness warnings
- Replace manual loops with .assign(x.begin(), x.end()) in RealVector
- Use static_cast<size_t>() for int to size_t comparisons
- Restructure ParameterVector constructor to avoid comparison warning

+++ inst/include/interface/rcpp/rcpp_objects/rcpp_growth.hpp
@@ -142,7 +142,7 @@ class EWAAGrowthInterface : public GrowthInterfaceBase {
   inline std::map<double, double> make_map(RealVector ages,
                                            RealVector weights) {
     std::map<double, double> mymap;
-    for (uint32_t i = 0; i < ages.size(); i++) {
+    for (size_t i = 0; i < ages.size(); i++) {
       mymap.insert(std::pair<double, double>(ages[i], weights[i]));
     }
     return mymap;
